# gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A gateway to our web-service",
        "title": "Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/auth/req_DH_params/": {
            "post": {
                "description": "Request DH Params from the server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Request DH Params",
                "parameters": [
                    {
                        "description": "Request data",
                        "name": "dh_params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.req_dh_params"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/req_pq/": {
            "post": {
                "description": "Request PQ from the server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Request PQ",
                "parameters": [
                    {
                        "description": "Request data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.req_pq_send_data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/biz/WithInjection/": {
            "post": {
                "description": "Get users from the server with SQL injection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get Users with Injection",
                "parameters": [
                    {
                        "description": "Request data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.get_users_injection_biz"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/biz/getUsers/": {
            "post": {
                "description": "Get users from the server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get Users",
                "parameters": [
                    {
                        "description": "Request data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.get_users_biz"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
        "main.get_users_biz": {
            "type": "object",
            "properties": {
                "auth_KEY": {
                    "type": "string"
                },
                "message_ID": {
                    "type": "integer"
                },
                "user_ID": {
                    "type": "string"
                }
            }
        },
        "main.get_users_injection_biz": {
            "type": "object",
            "properties": {
                "auth_KEY": {
                    "type": "string"
                },
                "message_ID": {
                    "type": "integer"
                },
                "user_ID": {
                    "type": "string"
                }
            }
        },
        "main.req_dh_params": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "integer"
                },
                "nonce": {
                    "type": "string"
                },
                "server_NONCE": {
                    "type": "string"
                }
            }
        },
        "main.req_pq_send_data": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}